requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
yfinance==0.2.28
schedule==1.2.0
pandas
matplotlib
certifi==2023.7.22
requests==2.31.0
